Bug 1257742 - Part 3: Support using floating point in sizes descriptor; r=jdm
MozReview-Commit-ID: ESTk2x2VqnD

UltraBlame original commit: 26bdf50a2e4507f5aca79d58b8d742aee8cd78e2

+++ dom/base/ResponsiveImageSelector.h
@@ -102,7 +102,7 @@ private:
   
   
   
-  bool ComputeFinalWidthForCurrentViewport(int32_t *aWidth);
+  bool ComputeFinalWidthForCurrentViewport(double* aWidth);
 
   nsCOMPtr<nsINode> mOwnerNode;
   
@@ -152,7 +152,7 @@ public:
   double Density(ResponsiveImageSelector *aSelector) const;
   
   
-  double Density(int32_t aMatchingWidth) const;
+  double Density(double aMatchingWidth) const;
 
   
   bool IsComputedFromWidth() const;
